group in same file CollisionGTComparison and CollisionGTStudy group in same file CorrelogramGTComparison and CorrelogramGTStudy
